Add button linking pseudoknot training collection to pseudoknot lab carousel slide (#338)
The API response for labs doesn't include any options for specifying training collections,
so until we add a feature for something like that, I've hardcoded a training collection
button into the lab slide that directly links to the correct collection. It will only
display for the pseudoknot lab now, but in the future we may want to add more options
for associating training collections with a given lab.